Refactoring wegen der Architektur
